# NotsHub.swagger.json
{"openapi":"3.1.0","info":{"title":"NotesPitara API","description":"API documentation for the NotesPitara application","contact":{"name":"Support","email":"support@notespitara.com"},"license":{"name":"NotesPitara License"},"version":"1.0"},"servers":[{"url":"http://localhost:5000","description":"Local Environment"},{"url":"https://notespitara.com","description":"Production Environment"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Semesters","description":"Endpoints for managing semesters"},{"name":"Universities","description":"Endpoints for managing universities"},{"name":"Subjects","description":"Endpoints for managing subjects"},{"name":"Branches","description":"Endpoints for managing university branches"},{"name":"Authentication","description":"Endpoints for user authentication and lifecycle management"},{"name":"Notes","description":"Endpoints for managing, uploading, and searching notes"},{"name":"Programs","description":"Endpoints for managing academic programs"}],"paths":{"/api/universities/{id}":{"put":{"tags":["Universities"],"operationId":"updateUniversity","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"request":{"$ref":"#/components/schemas/UniversityCreateRequest"},"logo":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["Universities"],"operationId":"deleteUniversity","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/subjects/{id}":{"get":{"tags":["Subjects"],"operationId":"getSubjectById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"put":{"tags":["Subjects"],"operationId":"updateSubject","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubjectCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["Subjects"],"operationId":"deleteSubject","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/semesters/{id}":{"put":{"tags":["Semesters"],"operationId":"updateSemester","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SemesterCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["Semesters"],"operationId":"deleteSemester","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/programs/{id}":{"get":{"tags":["Programs"],"operationId":"getProgramById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"put":{"tags":["Programs"],"operationId":"updateProgram","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProgramCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["Programs"],"operationId":"deleteProgram","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/notes/{id}":{"put":{"tags":["Notes"],"operationId":"updateNotes","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotesCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["Notes"],"operationId":"deleteNotes","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/notes/{id}/reject":{"put":{"tags":["Notes"],"operationId":"rejectNotes","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"rejectionNote","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/notes/{id}/approve":{"put":{"tags":["Notes"],"operationId":"approveNotes","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/branches/{id}":{"get":{"tags":["Branches"],"operationId":"getBranchById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"put":{"tags":["Branches"],"operationId":"updateBranch","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BranchCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"delete":{"tags":["Branches"],"operationId":"deleteBranch","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/universities":{"get":{"tags":["Universities"],"operationId":"getAllUniversities","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["Universities"],"operationId":"createUniversity","requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"request":{"$ref":"#/components/schemas/UniversityCreateRequest"},"logo":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/subjects":{"get":{"tags":["Subjects"],"operationId":"getSubjects","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["Subjects"],"operationId":"createSubject","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubjectCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/semesters":{"get":{"tags":["Semesters"],"operationId":"getAllSemesters","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["Semesters"],"operationId":"createSemester","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SemesterCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/programs":{"get":{"tags":["Programs"],"operationId":"getAllPrograms","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["Programs"],"operationId":"createProgram","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProgramCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/notes/upload":{"post":{"tags":["Notes"],"operationId":"uploadPdf","parameters":[{"name":"title","in":"query","required":true,"schema":{"type":"string"}},{"name":"description","in":"query","required":false,"schema":{"type":"string"}},{"name":"subjectId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/branches":{"get":{"tags":["Branches"],"operationId":"getAllBranches","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}},"post":{"tags":["Branches"],"operationId":"createBranch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BranchCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/verify-email-otp":{"post":{"tags":["Authentication"],"operationId":"verifyEmailOtp","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyEmailOtpRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/signup":{"post":{"tags":["Authentication"],"operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/signout":{"post":{"tags":["Authentication"],"operationId":"signoutUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/signin":{"post":{"tags":["Authentication"],"operationId":"authenticateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/reset-password":{"post":{"tags":["Authentication"],"operationId":"resetPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/resend-email-otp":{"post":{"tags":["Authentication"],"operationId":"resendEmailOtp","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResendEmailOtpRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/forgot-password":{"post":{"tags":["Authentication"],"operationId":"forgotPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/universities/{id}/programs":{"get":{"tags":["Universities"],"operationId":"getProgramsByUniversity","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/universities/slug/{slug}":{"get":{"tags":["Universities"],"summary":"Get a university by its slug","operationId":"getUniversityBySlug","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/subjects/{id}/notes":{"get":{"tags":["Subjects"],"operationId":"getNotesBySubject","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/subjects/slug/{slug}":{"get":{"tags":["Subjects"],"summary":"Get a subject by its slug","operationId":"getSubjectBySlug","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/semesters/{id}/subjects":{"get":{"tags":["Semesters"],"operationId":"getSubjectsBySemester","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/programs/{id}/branches":{"get":{"tags":["Programs"],"operationId":"getBranchesByProgram","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/programs/slug/{slug}":{"get":{"tags":["Programs"],"summary":"Get a program by its slug","operationId":"getProgramBySlug","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/notes":{"get":{"tags":["Notes"],"operationId":"getAllNotes","parameters":[{"name":"q","in":"query","required":false,"schema":{"type":"string"}},{"name":"query","in":"query","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/notes/{id}/download":{"get":{"tags":["Notes"],"operationId":"getDownloadLink","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/notes/slug/{slug}":{"get":{"tags":["Notes"],"summary":"Get notes by their slug","operationId":"getNotesBySlug","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/notes/search":{"get":{"tags":["Notes"],"operationId":"searchNotes","parameters":[{"name":"q","in":"query","required":false,"schema":{"type":"string"}},{"name":"query","in":"query","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/branches/{id}/semesters":{"get":{"tags":["Branches"],"operationId":"getSemestersByBranch","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/branches/slug/{slug}":{"get":{"tags":["Branches"],"summary":"Get a branch by its slug","operationId":"getBranchBySlug","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/users":{"get":{"tags":["Authentication"],"operationId":"getAllUsers","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/auth/username":{"get":{"tags":["Authentication"],"operationId":"currentUserName","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/auth/user":{"get":{"tags":["Authentication"],"operationId":"getUserDetails","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"UniversityCreateRequest":{"type":"object","properties":{"name":{"type":"string"},"code":{"type":"string","maxLength":20,"minLength":0},"description":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"logoUrl":{"type":"string"}},"required":["code","name"]},"SubjectCreateRequest":{"type":"object","properties":{"name":{"type":"string"},"code":{"type":"string"},"credits":{"type":"integer","format":"int32"},"syllabusUrl":{"type":"string"},"semesterId":{"type":"string","format":"uuid"}},"required":["code","credits","name","semesterId"]},"SemesterCreateRequest":{"type":"object","properties":{"number":{"type":"integer","format":"int32"},"branchId":{"type":"string","format":"uuid"}},"required":["branchId","number"]},"ProgramCreateRequest":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"type":{"type":"string"},"duration":{"type":"integer","format":"int32"},"universityId":{"type":"string","format":"uuid"}},"required":["duration","name","type","universityId"]},"NotesCreateRequest":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string","maxLength":5000,"minLength":0},"fileUrl":{"type":"string","maxLength":255,"minLength":0,"pattern":"^(https?://).+"},"fileKey":{"type":"string","maxLength":255,"minLength":0},"fileType":{"type":"string","pattern":"^(?i)PDF$"},"subjectId":{"type":"string","format":"uuid"}},"required":["fileKey","fileType","fileUrl","subjectId","title"]},"BranchCreateRequest":{"type":"object","properties":{"name":{"type":"string"},"code":{"type":"string","maxLength":20,"minLength":0},"programId":{"type":"string","format":"uuid"}},"required":["code","name","programId"]},"VerifyEmailOtpRequest":{"type":"object","properties":{"email":{"type":"string","maxLength":50,"minLength":0},"otp":{"type":"string","pattern":"^[0-9]{6}$"}},"required":["email","otp"]},"SignupRequest":{"type":"object","properties":{"username":{"type":"string","maxLength":20,"minLength":3,"pattern":"^[a-zA-Z0-9._-]+$"},"email":{"type":"string","maxLength":50,"minLength":0},"role":{"type":"array","items":{"type":"string"},"uniqueItems":true},"password":{"type":"string","maxLength":40,"minLength":6}},"required":["email","password","username"]},"LoginRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["password","username"]},"ResetPasswordRequest":{"type":"object","properties":{"token":{"type":"string"},"newPassword":{"type":"string","maxLength":120,"minLength":6}},"required":["newPassword","token"]},"ResendEmailOtpRequest":{"type":"object","properties":{"email":{"type":"string","maxLength":50,"minLength":0}},"required":["email"]},"ForgotPasswordRequest":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]}},"securitySchemes":{"bearerAuth":{"type":"http","description":"JWT authentication","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}